refactor: relocate AutoFacotory to seperate file
