+++ src/general_manager/factory/factories.py
@@ -1,159 +1,15 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, Type, Callable, Union, Any, TypeVar, Literal, cast
+from typing import Any, cast
 from factory.declarations import LazyFunction
 from factory.faker import Faker
 import exrex  # type: ignore
 from django.db import models
 from django.core.validators import RegexValidator
-from factory.django import DjangoModelFactory
 import random
 from decimal import Decimal
 from general_manager.measurement.measurement import Measurement
 from general_manager.measurement.measurementField import MeasurementField
 from datetime import date, datetime, time, timezone
-
-if TYPE_CHECKING:
-    from general_manager.interface.databaseInterface import (
-        DBBasedInterface,
-    )
-
-modelsModel = TypeVar("modelsModel", bound=models.Model)
-
-
-class AutoFactory(DjangoModelFactory[modelsModel]):
-    """
-    A factory class that automatically generates values for model fields,
-    including handling of unique fields and constraints.
-    """
-
-    interface: Type[DBBasedInterface]
-    _adjustmentMethod: (
-        Callable[..., Union[dict[str, Any], list[dict[str, Any]]]] | None
-    ) = None
-
-    @classmethod
-    def _generate(
-        cls, strategy: Literal["build", "create"], params: dict[str, Any]
-    ) -> models.Model | list[models.Model]:
-        cls._original_params = params
-        model = getattr(cls._meta, "model")
-        if not issubclass(model, models.Model):
-            raise ValueError("Model must be a type")
-        field_name_list, to_ignore_list = cls.interface.handleCustomFields(model)
-
-        fields = [
-            field
-            for field in model._meta.get_fields()
-            if field.name not in to_ignore_list
-        ]
-        special_fields: list[models.Field[Any, Any]] = [
-            getattr(model, field_name) for field_name in field_name_list
-        ]
-        pre_declations = getattr(cls._meta, "pre_declarations", [])
-        post_declarations = getattr(cls._meta, "post_declarations", [])
-        declared_fields: set[str] = set(pre_declations) | set(post_declarations)
-
-        field_list: list[models.Field[Any, Any] | models.ForeignObjectRel] = [
-            *fields,
-            *special_fields,
-        ]
-
-        for field in field_list:
-            if field.name in [*params, *declared_fields]:
-                continue  # Skip fields that are already set
-            if isinstance(field, models.AutoField) or field.auto_created:
-                continue  # Skip auto fields
-            params[field.name] = getFieldValue(field)
-
-        obj: list[models.Model] | models.Model = super()._generate(strategy, params)
-        if isinstance(obj, list):
-            for item in obj:  # type: ignore
-                if not isinstance(item, models.Model):
-                    raise ValueError("Model must be a type")
-                cls._handleManyToManyFieldsAfterCreation(item, params)
-        else:
-            cls._handleManyToManyFieldsAfterCreation(obj, params)
-        return obj
-
-    @classmethod
-    def _handleManyToManyFieldsAfterCreation(
-        cls, obj: models.Model, attrs: dict[str, Any]
-    ) -> None:
-        for field in obj._meta.many_to_many:
-            if field.name in attrs:
-                m2m_values = attrs[field.name]
-            else:
-                m2m_values = getManyToManyFieldValue(field)
-            if m2m_values:
-                getattr(obj, field.name).set(m2m_values)
-
-    @classmethod
-    def _adjust_kwargs(cls, **kwargs: dict[str, Any]) -> dict[str, Any]:
-        # Remove ManyToMany fields from kwargs before object creation
-        model: Type[models.Model] = getattr(cls._meta, "model")
-        m2m_fields = {field.name for field in model._meta.many_to_many}
-        for field_name in m2m_fields:
-            kwargs.pop(field_name, None)
-        return kwargs
-
-    @classmethod
-    def _create(
-        cls, model_class: Type[models.Model], *args: list[Any], **kwargs: dict[str, Any]
-    ) -> models.Model | list[models.Model]:
-        kwargs = cls._adjust_kwargs(**kwargs)
-        if cls._adjustmentMethod is not None:
-            return cls.__createWithGenerateFunc(strategy=True, params=kwargs)
-        return cls._modelCreation(model_class, **kwargs)
-
-    @classmethod
-    def _build(
-        cls, model_class: Type[models.Model], *args: list[Any], **kwargs: dict[str, Any]
-    ) -> models.Model | list[models.Model]:
-        kwargs = cls._adjust_kwargs(**kwargs)
-        if cls._adjustmentMethod is not None:
-            return cls.__createWithGenerateFunc(strategy=False, params=kwargs)
-        return cls._modelBuilding(model_class, **kwargs)
-
-    @classmethod
-    def _modelCreation(
-        cls, model_class: Type[models.Model], **kwargs: dict[str, Any]
-    ) -> models.Model:
-        obj = model_class()
-        for field, value in kwargs.items():
-            setattr(obj, field, value)
-        obj.full_clean()
-        obj.save()
-        return obj
-
-    @classmethod
-    def _modelBuilding(
-        cls, model_class: Type[models.Model], **kwargs: dict[str, Any]
-    ) -> models.Model:
-        obj = model_class()
-        for field, value in kwargs.items():
-            setattr(obj, field, value)
-        return obj
-
-    @classmethod
-    def __createWithGenerateFunc(
-        cls, strategy: bool, params: dict[str, Any]
-    ) -> models.Model | list[models.Model]:
-        model_cls = getattr(cls._meta, "model")
-        if cls._adjustmentMethod is None:
-            raise ValueError("generate_func is not defined")
-        records = cls._adjustmentMethod(**params)
-        if isinstance(records, dict):
-            if strategy:
-                return cls._modelCreation(model_cls, **records)
-            return cls._modelBuilding(model_cls, **records)
-
-        created_objects: list[models.Model] = []
-        for record in records:
-            if strategy:
-                created_objects.append(cls._modelCreation(model_cls, **record))
-            else:
-                created_objects.append(cls._modelBuilding(model_cls, **record))
-        return created_objects
 
 
 def getFieldValue(field: models.Field[Any, Any] | models.ForeignObjectRel) -> object:
